fs_cache:  - solucionado el bug que impedía añadir elementos a la cache. modelos:  - sustituidas todas las llamadas a intval() por $this->intval(). Esta función devuelve NULL en caso de encontrar un NULL, al contrario que la anterior, que devuelve 0. general_nuevo_albaran, tpv_recambios, tpv_yamyam:  - se muestra un error si no se encuentra una dirección asociada al cliente.
git-svn-id: http://facturascripts.googlecode.com/svn/branches/0.9@377 fe615fe8-962f-0410-811c-9ded47c8d053

+++ model/albaran_proveedor.php
@@ -44,8 +44,8 @@ class linea_albaran_proveedor extends fs_model
       parent::__construct('lineasalbaranesprov');
       if($l)
       {
-         $this->idlinea = intval($l['idlinea']);
-         $this->idalbaran = intval($l['idalbaran']);
+         $this->idlinea = $this->intval($l['idlinea']);
+         $this->idalbaran = $this->intval($l['idalbaran']);
          $this->referencia = $l['referencia'];
          $this->descripcion = $l['descripcion'];
          $this->cantidad = floatval($l['cantidad']);
@@ -231,8 +231,8 @@ class albaran_proveedor extends fs_model
       parent::__construct('albaranesprov');
       if($a)
       {
-         $this->idalbaran = intval($a['idalbaran']);
-         $this->idfactura = intval($a['idfactura']);
+         $this->idalbaran = $this->intval($a['idalbaran']);
+         $this->idfactura = $this->intval($a['idfactura']);
          $this->codigo = $a['codigo'];
          $this->numero = $a['numero'];
          $this->numproveedor = $a['numproveedor'];
@@ -321,17 +321,17 @@ class albaran_proveedor extends fs_model
    
    public function factura_url()
    {
-      if( !$this->ptefactura )
+      if( $this->ptefactura )
+         return '#';
+      else
       {
          $fac = new factura_proveedor();
          $fac = $fac->get($this->idfactura);
          if($fac)
             return $fac->url();
          else
-            return $this->url();
+            return '#';
       }
-      else
-         return $this->url();
    }
    
    public function agente_url()
@@ -459,7 +459,13 @@ class albaran_proveedor extends fs_model
    
    public function delete()
    {
-      return $this->db->exec("DELETE FROM ".$this->table_name." WHERE idalbaran = '".$this->idalbaran."';");
+      if($this->ptefactura AND (is_null($this->idfactura) OR $this->idfactura == 0))
+         return $this->db->exec("DELETE FROM ".$this->table_name." WHERE idalbaran = '".$this->idalbaran."';");
+      else
+      {
+         $this->new_error_msg("Este albarán está vinculado a una factura.");
+         return FALSE;
+      }
    }
    
    public function all($offset=0)
